==> notebooks/03b EDA - automatic EDA.ipynb <==
### Imports

import pandas as pd

from ydata_profiling import ProfileReport


file_path =  "../data/rents_returns_by_bike_Dresden_01-03_09-10.24.csv" 

df = pd.read_csv(file_path, parse_dates = ["datetime", "date", "datetime_hour"])

df.head()

df.info()

df.describe()

df.tail()

report = ProfileReport(df)


report.to_file("../data/output/ydata_profiling of df.html")
