==> notebooks/04c Browse.AI continued.ipynb <==
# Imports and settings

#from elasticsearch_dsl import Search, Q, Range, connections
#from elasticsearch.client import Elasticsearch
from tqdm.notebook import tqdm
from datetime import datetime
import pandas as pd
import matplotlib.pyplot as plt
import os
import json
import re
from openai import OpenAI
import numpy as np

DATA_FOLDER = "../data/events/browse_ai/" 

pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", 20)
pd.set_option('display.max_colwidth', 300)

df = pd.read_pickle(f"{DATA_FOLDER}df_browse_ai before description 2024-12-19_18-27-21.pkl")

len(df)

df.head()

df.replace("nan", np.nan, inplace=True)

extracted_columns = [col for col in df.columns if col.startswith("extr")]
required_cols = ["Event_URL", "isOnMultipleDays"]
required_cols.extend(extracted_columns)
required_cols_except_start_date = required_cols.copy()
required_cols_except_start_date.remove("extrStartDate")

df['count_notna'] = df[required_cols].notna().sum(axis=1)

df = df.loc[df.groupby('Event_URL')['count_notna'].idxmax()]

len(df)

### manual data replacements

flt = df["Event_description"].str.contains("Caspar David Friedrich. Wo alles begann", na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp

df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrStreetHouseNumber"]="Tzschirnerpl. 2"
df.loc[id, "extrMultipleDaysDetails"] = """Freitag-Montag, Mittwoch 10-18 Uhr Donnerstag 10-21 Uhr Dienstag geschlossen"""



flt = df["Event_description"].str.contains("Napoleon mit seiner Grande Armée marschierte am 24.", na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-27"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Militärhistorisches Museum der Bundeswehr"
df.loc[id, "extrStreetHouseNumber"]="Olbrichtpl. 2"
df.loc[id, "extrLocationDetails"]="Olbrichtpl. 2, 01099 Dresden"
df.loc[id, "extrMultipleDaysDetails"] = """Montag 10 - 21 Uhr Dienstag, Donnerstag, Freitag, Samstag, Sonntag 10-18 Uhr Mittwoch geschlossen"""


flt = df["Event_description"].str.contains("Wer entscheidet eigentlich, wer dazugehört", na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Robotron-Kantine in Dresden"
df.loc[id, "extrStreetHouseNumber"]="Zinzendorfstraße 5"
df.loc[id, "extrLocationDetails"]="Zinzendorfstraße 5, 01069 Dresden"
df.loc[id, "extrMultipleDaysDetails"] = """Mi – Fr 16 – 19 Uhr Sa & So und Feiertag 12 -18 Uhr"""


flt = df["Event_description"].str.contains("Eine Ausstellung zeitgenössischer Kunst zu den Auswirkungen rassistischer", na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Robotron-Kantine in Dresden"
df.loc[id, "extrStreetHouseNumber"]="Zinzendorfstraße 5"
df.loc[id, "extrLocationDetails"]="Zinzendorfstraße 5, 01069 Dresden"
df.loc[id, "extrMultipleDaysDetails"] = """Mi – Fr 16 – 19 Uhr Sa & So und Feiertag 12 -18 Uhr"""


flt = df["Event_description"].str.contains('Unter dem Motto "Über die Natur" findet zum vierten Mal ', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Schloss Pillnitz"
df.loc[id, "extrLocationDetails"]="August-Böckstiegel-Straße 2, 01326 Dresden"
df.loc[id, "extrStreetHouseNumber"]="August-Böckstiegel-Straße 2"
df.loc[id, "extrMultipleDaysDetails"] = """täglich 10-17 Uhr, Montag geschlossen"""


flt = df["Event_description"].str.contains('Mikroschnitzereien auf Streichhölzern oder in Walnussschalen, Modelle und Puppenspielzeug', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-20"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationDetails"]="Köpckestraße 1, 01097 Dresden"
df.loc[id, "extrStreetHouseNumber"]="Köpckestraße 1"
df.loc[id, "extrMultipleDaysDetails"] = """täglich 10-18 Uhr, Montag geschlossen"""


flt = df["Event_description"].str.contains('Dem langjährigen Professor für Monumentalmalerei widmet sich die neue Sonderausstellung', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Schloss Pillnitz"
df.loc[id, "extrLocationDetails"]="August-Böckstiegel-Straße 2, 01326 Dresden"
df.loc[id, "extrStreetHouseNumber"]="August-Böckstiegel-Straße 2"
df.loc[id, "extrMultipleDaysDetails"] = """Dienstag bis Sonntag 10 bis 17 Uhr"""


flt = df["Event_description"].str.contains('Tonband- und Kassettengeräte kam man', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-19"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]=" Städtische Galerie Dresden"
df.loc[id, "extrMultipleDaysDetails"] = """Eröffnung: DO 19.09. 19.00 Uhr DI bis SO 10.00 bis 18.00 Uhr FR 10.00 bis 19.00 Uhr"""


flt = df["Event_description"].str.contains('100 Jahre KUNSTAUSSTELLUNG KÜHL in Dresden.', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]=" Kunstausstellung Kühl Dresden"
df.loc[id, "extrStreetHouseNumber"]="Nordstraße 5"
df.loc[id, "extrMultipleDaysDetails"] = """Mi bis Fr 11 –19 + Sa 11 –16 Uhr"""


flt = df["Event_description"].str.contains('Auf die Elbe, fertig, los!', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Blaues Wunder"
df.loc[id, "extrStartTime"] = "11"
df.loc[id, "extrEndTime"] = "15"

flt = df["Event_description"].str.contains('Das Highlight für Kinder ist die Parkeisenbahn.', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-26"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Parkeisenbahn Dresden"
df.loc[id, "extrLocationDetails"]="01219 Dresden-Altstadt"
df.loc[id, "extrMultipleDaysDetails"] = """Mittwoch bis Freitag: 13:00 bis 18:00 Uhr Samstag, Sonntag, Feiertag: 10:30 bis 18:00 Uhr"""

flt = df["Event_description"].str.contains('Süße, halb tote Katze verschwunden! Ihre Spur', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Technische Sammlungen Dresden"
df.loc[id, "extrLocationDetails"]="01277 Dresden Junghannsstraße 1-3"
df.loc[id, "extrStreetHouseNumber"]="Junghannsstraße 1-3"
df.loc[id, "extrMultipleDaysDetails"] = """DI bis FR 9.00 bis 17.00 Uhr SA/SO 10.00 bis 18.00 Uhr"""

flt = df["Event_description"].str.contains('prägenden Periode in der Geschichte des DHMD und stellt sie in den Kontext der wohl einschneidendsten Umbrüche der', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationName"]="Hygienemuseum"
df.loc[id, "extrLocationDetails"]="01069 Dresden Lingnerplatz 1"
df.loc[id, "extrStreetHouseNumber"]="Lingnerplatz 1"
df.loc[id, "extrMultipleDaysDetails"] = """Dienstag bis Sonntag, Feiertage: 10 bis 18 Uhr"""

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Kinderbiennale-Planet-Utopia-Dresden-114311.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationDetails"]="01097 Dresden Palaisplatz 11"
df.loc[id, "extrStreetHouseNumber"]="Palaisplatz 11"
df.loc[id, "extrMultipleDaysDetails"] = "Diensatg bis Sonntag 10 bis 18 Uhr"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Wiederentdeckt-Dresdner-Malerinnen-Der-Romantik-Dresden-116639.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationDetails"]="01097 Dresden Hauptstr. 13"
df.loc[id, "extrMultipleDaysDetails"] = "MI bis FR 10.00 bis 17.00 Uhr SA/SO 12.00 bis 17.00 Uhr"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Von-Wanderlust-Und-Reisefrust', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-27"
df.loc[id, "extrCity"]="Dresden"
df.loc[id, "extrLocationDetails"]="01099 Dresden Nordstraße 28"
df.loc[id, "extrMultipleDaysDetails"] = "MI bis SO 12.00 bis 17.00 Uhr"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Ausstellung-Home-Sweet-Home-Danny-Hermann-Radebeul-117962.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-07"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrMultipleDaysDetails"] = "Samstag & Sonntag 14-18 Uhr"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Sonderausstellung-Andreas-Braeunsdorf-Malerei-Dresden-116638.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-07"
df.loc[id, "extrEndDate"] = "2024-09-29"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Udo-Zimmermann-Ein-Moderner-Romantiker-Dresden-114359.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-09-29"
df.loc[id, "extrMultipleDaysDetails"] = "MI bis SO 12.00 bis 17.00 Uhr"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Trickfilm-Plakat-Kunst-Arbeiten-Fuer-Den-Ddr-Filmverleih-113199.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Panorama-Amazonien-Im-Panometer-Dresden-Dresden-113732.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-31"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Zeitfenster-Spurensuche-Zur-Industriegeschichte-Im-Ernemannbau-Dresden-113361.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-20"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Physics-Of-Life-Physik-Des-Lebens-Dresden-109648.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-09-01"
df.loc[id, "extrEndDate"] = "2024-10-27"

flt = df["Event_URL"].str.contains('https://sachsen-net.com/veranstaltungen/heute/Dresdner-Herbstvolksfest-Der-Freizeitpark-Dresden-119283.html', na=False)
id = max(df.index) + 1
print(id)
tmp = df[flt]
assert len(df[flt])>0 
df = df[~flt]
tmp = tmp.iloc[0]
print(tmp)
df.loc[id]=tmp
df.loc[id, "isOnMultipleDays"] = True
df.loc[id, "extrStartDate"] = "2024-10-03"
df.loc[id, "extrEndDate"] = "2024-10-31"
df.loc[id, "extrMultipleDaysDetails"] = """ Montag bis Donnerstag 15:00 - 22:00 Uhr
Freitag 15:00 - 23:00 Uhr
Samstag 14:00 - 23:00 Uhr
Sonntag 14:00 - 22:00 Uhr
SCHLIESSTAGE 21.10., 22.10., 28.10., 29.10.2024"""
df.loc[id, "extrLocationName"]="Volksfestgelände Pieschener Allee"
df.loc[id, "extrLocationDetails"]="01067 Dresden Pieschener Allee 1"

### extra scraped info from browse ai

len(df)

df.loc[:, "count_Event_URL"]= df.Event_URL.map(df.Event_URL.value_counts())

df.query("count_Event_URL>1")[["isOnMultipleDays", "Event_URL"]].value_counts()

len(df)

# df_urls.to_csv(f"{DATA_FOLDER}urls_to_check.csv", index=False)

df_scraped = pd.read_csv(f"{DATA_FOLDER}urls_to_check_scraped.csv")

df.columns

for index, row in df_scraped.iterrows():
    url = row["Origin URL"]
    dates_scraped = row["dates"]

    location_details = row["location_details"]
    location_name = row["location_name"]
    print(location_details)
    print(location_name)

    if "Flughafen" in dates_scraped:
        dates_possibly_with_time=dates_scraped.split(",")[0]
        start_time="15"
        end_time="16,5"
    elif "Dienstag" in dates_scraped:
        dates_possibly_with_time=dates_scraped.split(",")[0]
    else:
        split = dates_scraped.split(" - ")
        assert len(split)==2, split
        dates_possibly_with_time = split[0]

    if "," in dates_possibly_with_time:
        dates, times = dates_possibly_with_time.split(",")[0], dates_possibly_with_time.split(",")[1]
        has_time=True
    else:
        dates = dates_possibly_with_time
        has_time=False


    start_date = pd.to_datetime(dates.split(" bis ")[0], dayfirst=True).strftime("%Y-%m-%d")
    end_date = pd.to_datetime(dates.split(" bis ")[1], dayfirst=True).strftime("%Y-%m-%d")
    flt = df["Event_URL"]==url

    assert len(df.loc[flt]) > 0, url

    df.loc[flt, "extrStartDate"] = start_date
    df.loc[flt, "extrEndDate"] = end_date
    df.loc[flt, "isOnMultipleDays"] = True
    df.loc[flt, "extrLocationName"] = location_name
    df.loc[flt, "extrLocationDetails"] = location_details

    if has_time:
        split=times.split(" bis ")
        assert len(split)==2, split
        start_time = split[0].replace(".00", "").strip()
        end_time = split[1].replace(" Uhr", "").replace(".00", "").strip()
        df.loc[flt, "extrStartTime"] = start_time
        df.loc[flt, "extrEndTime"] = end_time

    
    
    
    
    

df_scraped.head()

len(df)

# check 1 scraped data
flt = df.Event_URL=="https://sachsen-net.com/veranstaltungen/heute/Sonderausstellung-Wilhelm-Mueller-Dresden-118890.html"
df.loc[flt, required_cols]

flt = df.extrLocationDetails == "August-Böckstiegel-Straße 2, 01326 Dresden"
df.loc[flt, required_cols]

### extr start time from  Start_date_and_end_date

def change_PM_hour_minute_to_my_format(s):
    spl = s.split(":")
    hour= int(spl[0])
    if hour<12:
        hour=hour+12
    hour=str(hour)
    if spl[1]=="15":
        minute=""
    elif spl[1]=="30":
        minute=",5"
    else:
        minute=""

    return "".join([hour, minute])
        

flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna())  &  (~ (df.Start_date_and_end_date.isna()))
print(len(df[flt]))
df.loc[flt, "Start_date_and_end_date"].unique()


df.columns

flt = df.Start_date_and_end_date.str.endswith(" PM\xa0Time", na=False)
values = df.loc[flt, "Start_date_and_end_date"].str.replace(" PM\xa0Time", "").apply(change_PM_hour_minute_to_my_format)
df.loc[flt, "extrStartTime"]=values

flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna())  &  (~ (df.Start_date_and_end_date.isna()))
print(len(df[flt]))
df.loc[flt, "Start_date_and_end_date"].unique()


flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna())  &  (~ (df.Start_date_and_end_date.isna())) & (df.Start_date_and_end_date=="10:30 AM\xa0Time")
print(len(df[flt]))
df.loc[flt, "extrStartTime"]="10,5"
values = df.loc[flt, "Start_date_and_end_date"].str.replace(" PM\xa0Time", "").apply(change_PM_hour_minute_to_my_format)


flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna())  &  (~ (df.Start_date_and_end_date.isna()))
print(len(df[flt]))
df.loc[flt, "Start_date_and_end_date"].unique()


flt = df.Start_date_and_end_date.str.endswith(" Uhr", na=False)
values = df.loc[flt, "Start_date_and_end_date"]
values.str.split("\n").apply(lambda ls: ls[-1]).str.replace(" Uhr","").str.replace(":30", ",5").str.replace(":00", "")
df.loc[flt, "extrStartTime"]=values

flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna()) 
print(len(df[flt]))
df.loc[flt, "Event_URL"].apply(lambda url: url[:30]).value_counts()


flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna()) & (df.Event_URL.str.startswith("https://www.veranstaltung-bade") )
df.loc[flt]

id = 10430
df.loc[id, "extrStartTime"]="20"
id = 10468
df.loc[id, "extrStartTime"]="20"
id = 25851
df.loc[id, "extrStartTime"]="18,5"
id = 25832
df.loc[id, "extrStartTime"]="18,5"

### additionally scrap start time for 170 sachsen-net events

flt = (df.isOnMultipleDays==False ) &  (df.extrStartTime.isna())  & (df.Event_URL.str.contains("sachsen-net"))
df.loc[flt, "Event_URL"].drop_duplicates().to_csv(f"{DATA_FOLDER}sachsen-net-single-events-to-extract-start-time.csv", index=False)

flt = df.Event_URL.str.contains("https://sachsen-net.com/veranstaltungen/morgen/Erlebnisausstellung-Die-Spiele-Des-Koenigs-Mit-Ringrennkarussell-Dresden")
df.loc[flt, "isOnMultipleDays"]=True
df.loc[flt, "extrStartDate"]="2024-09-01"
df.loc[flt, "extrEndDate"]="2024-09-29"
df.loc[flt, "extrStartTime"]="10"
df.loc[flt, "extrEndTime"]="17"

df_scraped = pd.read_csv(f"{DATA_FOLDER}sachsen-net-single-events-extracted-start-time.csv")

df_scraped.tail()

def change_sachsen_net_hm_to_standard_format(s):
    spl = s.split(":")
    hour= int(spl[0])
    if spl[1]=="15":
        minute=""
    elif spl[1]=="30":
        minute=",5"
    elif spl[1]=="45":
        minute=""
        hour+=1
    else: 
        minute=""
    return "".join([str(hour), minute])
        

for index, row in df_scraped.iterrows():
    url = row["Origin URL"]
    print(url)
    dates_scraped = row["dates"]
    print(dates_scraped)
    try:
        end_time = change_sachsen_net_hm_to_standard_format(row["EndTime"].replace("Uhr", "").strip())
    except KeyError:
        end_time = None
    location_details = row["location_details"]
    location_name = row["location_name"]

    
    print(location_details)
    print(location_name)

    split = dates_scraped.split(",")
    start_time = change_sachsen_net_hm_to_standard_format(split[-1].replace("Uhr", "").strip())

    print(f"{start_time=}")
    print(f"{end_time=}")

    flt = df["Event_URL"]==url
    assert len(df.loc[flt]) > 0, url

    df.loc[flt, "extrStartTime"] = start_time
    df.loc[flt, "extrEndTime"] = start_time
    df.loc[flt, "extrEndDate"] = None
    df.loc[flt, "isOnMultipleDays"] = False
    df.loc[flt, "extrLocationName"] = location_name
    df.loc[flt, "extrLocationDetails"] = location_details



    
    
    
    
    

# estimate the readiness of the dataset

len(df)

extracted_columns = []
for col in df.columns:
    if col.startswith("extr"):
        extracted_columns.append(col)
    
extracted_columns.append("isOnMultipleDays")

df.extrEndDate.isna().value_counts()

flt = df.isOnMultipleDays==False
print("total single-day events")
print(len(df[flt]))
for col in extracted_columns:
    flt_2 = df[col].isna()
    print(col+ " has missing rows:")
    print(len(df[flt&flt_2]))
    


df.loc[flt, extracted_columns]

## save

exclude_columns= ['Date (GMT)', 'Job Link']

subset_columns = [col for col in df.columns if col not in exclude_columns]

df = df.drop_duplicates(subset=subset_columns)

len(df)

df = df.dropna(how="all", axis=1)

len(df)

# df.to_excel(f"{DATA_FOLDER}df_browse_ai {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.xlsx")

df.to_csv(f"{DATA_FOLDER}df_browse_ai {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.csv")

df.to_pickle(f"{DATA_FOLDER}df_browse_ai {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.pkl")
