==> notebooks/05a Unite all events into 1 dataset.ipynb <==
# setup

#from elasticsearch_dsl import Search, Q, Range, connections
#from elasticsearch.client import Elasticsearch
from tqdm.notebook import tqdm
from datetime import datetime
import pandas as pd
import matplotlib.pyplot as plt
import os
import json
import re
from openai import OpenAI
import numpy as np

EVENTS_FOLDER = "../data/events/"

pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", 20)
pd.set_option('display.max_colwidth', None)

# browse_ai

df_browse_ai = pd.read_pickle(f"{EVENTS_FOLDER}/browse_ai/df_browse_ai final 2025-01-07_18-08-29.pkl")

df_browse_ai.rename(columns={"Event_category":"EventCategory"}, inplace=True)

df_browse_ai.rename(columns={col: col.replace("extr", "") for col in df_browse_ai.columns}, inplace=True)

df_browse_ai.columns

df_browse_ai["lat"] = None
df_browse_ai["lon"] = None
df_browse_ai["Street"] = None
df_browse_ai["HouseNumber"] = None
df_browse_ai['StartTimeIsAssumed']=False

df_browse_ai.drop(columns="MultipleDaysDetails", inplace=True)

df_browse_ai.columns

# Gema

df_gema = pd.read_pickle(f"{EVENTS_FOLDER}/gema/df_gema 2024-12-19_12-44-50.pkl")

flt = df_gema.statusDescription=="Storniert"
df_gema = df_gema[~flt]

df_gema.haeufigkeit.value_counts()

df_gema['Source']="Gema"
df_gema['StartTimeIsAssumed']=True

for col in ["LocationDetails", "StartTime", "EndTime"]:
    df_gema[col] = None


df_gema.rename(columns={"gemaNo_address_coordinate_lat":"lat", 
                        "gemaNo_address_coordinate_lon": "lon",
                        "gemaNo_address_streetHouseNumber": "StreetHouseNumber",
                        "gemaNo_address_city": "City",
                        "gemaNfNr": "origin_id",
                        "startDate": "StartDate",
                        "gemaNo_address_street": "Street",
                        "gemaNo_address_houseNumber": "HouseNumber",
                        "gemaNo_categoryType": "EventCategory",
                        "gemaNo_name": "LocationName",
                        "endDate":"EndDate"
                        }, inplace=True)

pd.to_datetime(df_gema['EndDate'].dropna())

flt = df_gema.EndDate == df_gema.StartDate
df_gema.loc[flt, "EndDate"] = None

df_gema.StartDate = pd.to_datetime(df_gema.StartDate)
df_gema.EndDate = pd.to_datetime(df_gema.EndDate)

df_gema["duration"] = (df_gema.EndDate - df_gema.StartDate).dt.days

# df_gema.statusDescription.value_counts()

# pd.set_option("display.max_rows", None)
# display(df_gema.loc[df_gema["duration"] > 60, "duration"].value_counts().sort_index())
# pd.set_option("display.max_rows", 20)


#remove suspiciously long events
flt = (df_gema["duration"] <= 60) | (df_gema["duration"].isnull())
df_gema = df_gema[flt]

len(df_gema)

# pd.set_option("display.max_rows", None)
# display(df_gema.loc[:, "duration"].value_counts().sort_index())
# pd.set_option("display.max_rows", 20)


# df_gema[df_gema.duration>=59]

# df_gema.description.value_counts()

# df_gema.EventCategory.value_counts().index

#df_gema.statusDescription.value_counts()

df_gema["isOnMultipleDays"]=False
df_gema.loc[df_gema["duration"]>0, "isOnMultipleDays"]=True

df_gema.loc[df_gema.isOnMultipleDays==True, [ "StartDate", "EndDate", "duration"]]

for col in df_browse_ai.columns:
    if col not in df_gema.columns:
        print(col)

# for col in df_gema.columns:
#     if  "name" in col.lower():
#         print(col)

# df_gema.gemaNo_name.value_counts()

# df_gema.gemaNo_branche.value_counts()

# df_gema.gemaNo_branchensegment.value_counts()

# df_gema.gemaNo_categoryType.value_counts()

# df_browse_ai.EventCategory.value_counts()

# assume multiple events repeat every day. 

# flt = ~df_gema.EndDate.isna()
# df_gema.loc[flt, "isOnMultipleDays"].value_counts()

# flt = (~df_gema.EndDate.isna()) & (df_gema.isOnMultipleDays==False) 
# df_gema.loc[flt, "EndDate"].value_counts()

# flt = (~df_gema.endDate.isna()) & (df_gema.isOnMultipleDays==False) # & (df_gema.StartDate!=df_gema.endDate)
# len(df_gema.loc[flt])

# df_gema.duration.value_counts()

# df_gema.lat.isna().value_counts()

# df_gema.endDate.dropna()

df_gema.columns

cols_to_keep = df_browse_ai.columns.tolist()
df_gema = df_gema[cols_to_keep]

df_gema_to_add = df_gema.loc[df_gema.isOnMultipleDays==False].copy()

df_tmp = df_gema.loc[df_gema.isOnMultipleDays==True].copy()

len(df_tmp)

df_tmp

rows_to_append = []
for i, row in df_gema.loc[df_gema.isOnMultipleDays==True].iterrows():
    start_date = row["StartDate"]
    end_date = row["EndDate"]
    print(f"{start_date=}")
    print(f"{end_date=}")
    date_range = pd.date_range(start_date, end_date)
    
    for date in date_range: 
        row_to_append = row.copy()
        row_to_append["StartDate"] = date.strftime('%Y-%m-%d')
        rows_to_append.append(row_to_append)

pd.DataFrame(rows_to_append)

df_gema_final = pd.concat([df_gema_to_add, pd.DataFrame(rows_to_append)], axis=0)

df_gema_final

df_gema_final.to_pickle(f"{EVENTS_FOLDER}/gema/df_gema after expanding multiple days events {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.pkl")

df_gema_final.columns

# apify

# read a json file
with open(f"{EVENTS_FOLDER}apify/apify original DD FB.json") as f:
    apify = json.load(f)

rows_to_append = []
for element in apify:
    event = element['event']
    location = element['location']
    location_address = location['address']
    row_to_append = {}
    # row_to_append["StartDate"] = event['startDateTime']
    # row_to_append["StartTime"] = event['startDateTime']
    # row_to_append["EndTime"] = event['endDateTime']
    # row_to_append["EndDate"] = event['endDateTime']
    row_to_append["StartDateTime"] = event['startDateTime']
    row_to_append["EndDateTime"] = event['endDateTime']
    row_to_append["City"] = location_address['city']
    row_to_append["LocationDetails"] = location_address['rawAddress']
    row_to_append["LocationName"] = location['name']
    row_to_append["Street"] = location_address['street']
    row_to_append["HouseNumber"] = location_address['houseNumber']
    row_to_append['origin_id'] = event['sourceInformation']['uuid']
    row_to_append['EventCategory'] = location['category'] if event['category'] is None else event['category']
    row_to_append['Source'] = event['sourceInformation']['platform']
    row_to_append['EventDescription'] = event['description']
    rows_to_append.append(row_to_append)

df_apify = pd.DataFrame(rows_to_append)

flt = df_apify.StartDateTime == df_apify.EndDateTime
df_apify.loc[flt, "EndDateTime"] = None


for col in ['StreetHouseNumber', 'isOnMultipleDays', 'lat', 'lon', 'StartTimeIsAssumed', 'StartDate', 'EndDate', 'StartTime', 'EndTime']:
    df_apify[col] = None

for col in df_gema_final.columns:
    if col not in df_apify.columns:
        print(col)

df_apify.tail()

df_apify.Source = df_apify.Source.str.replace("https://www.", "",  regex=False)

# df_apify.EndDate.isna().value_counts()

# df_apify.StartDate.isna().value_counts()

# df_apify.StartDate.unique()

for col in ['StartDateTime', 'EndDateTime']:
    df_apify[col] = df_apify[col].astype(str).str.replace("Z", "").str.replace("+02:00", "", regex=False).str.replace("+01:00", "", regex=False)
    #df_apify[col] = pd.to_datetime(df_apify[col], errors="coerce")

df_apify.StartDateTime = pd.to_datetime(df_apify.StartDateTime, yearfirst=True)


df_apify.EndDateTime = pd.to_datetime(df_apify.EndDateTime, yearfirst=True, errors="coerce")


df_apify.StartDateTime.unique()

df_apify.EndDateTime.unique()

# flt = df_apify.EndDate.dt.year==2025
flt = df_apify.origin_id=="33a3e0d0-8141-11ef-aa2c-ed8353722086"
df_apify = df_apify[~flt]

# for col in ['StartDate', "EndDate"]: 
#     df_apify[col] = df_apify[col].dt.strftime('%Y-%m-%d')


# flt = df_apify.EndDate.isna()
# (df_apify.loc[~flt].EndTime-df_apify.loc[~flt].StartTime).sort_values()

df_apify['isOnMultipleDays']=False

# df_apify.StartTime = df_apify.StartTime.apply(lambda x: x.strftime('%H:%M'))

# df_apify.EndTime = df_apify.EndTime.apply(lambda x: x.strftime('%H:%M') if x is not pd.NaT else None)

# df_apify.EndTime.unique()

# df_apify.StartTime.unique()

df_apify.head()

# flt =(~(df_apify.StartDate.str.contains("\+02:00"))) & (~(df_apify.StartDate.str.contains("\+01:00")))
# df_apify.loc[flt, ['StartDate', 'EventDescription', 'Source']]

# apify[-1]

# apify[0]['location']

# apify[-1]['event']['description']

df_events = pd.concat([df_browse_ai, df_gema_final, df_apify], axis=0)

df_events.head()

df_events.to_pickle(f"{EVENTS_FOLDER}/df_events have to add meslis {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.pkl")
