==> notebooks/05b Unite all events into 1 dataset-continued.ipynb <==
# setup

#from elasticsearch_dsl import Search, Q, Range, connections
#from elasticsearch.client import Elasticsearch
from tqdm.notebook import tqdm
from datetime import datetime
import pandas as pd
import matplotlib.pyplot as plt
import os
import json
import re
from openai import OpenAI
import numpy as np
import ast

EVENTS_FOLDER = "../data/events/"

pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", 20)
pd.set_option('display.max_colwidth', None)

df_events = pd.read_pickle(f"{EVENTS_FOLDER}df_events have to add meslis 2025-01-07_22-23-01.pkl")

# shuffle the data rows 
df_events.sample(frac=1).reset_index(drop=True)

# meslis

df_meslis = pd.read_pickle(f"{EVENTS_FOLDER}meslis/df_meslis 2024-12-19_12-41-21.pkl")

df_meslis.isCancelled.value_counts()

flt = df_meslis.isCancelled==True
df_meslis = df_meslis[~flt]

### standardize columns

df_meslis.rename(columns = {
    "id": "origin_id",
    "startDate": "StartDateTime",
    "endDate": "EndDateTime",
    "mergedLocation_name": "LocationName",
    "mergedLocation_address_city": "City",
    "mergedLocation_address_street": "Street",
    "mergedLocation_address_houseNumber": "HouseNumber",
    "mergedLocation_address_streetHouseNumber": "StreetHouseNumber",
    "mergedLocation_coordinate_lat": "lat",
    "mergedLocation_coordinate_lon": "lon", 

    
}, inplace=True)

df_meslis.isDeleted.value_counts()

df_meslis = df_meslis[df_meslis.isDeleted == False]

# for col in df_meslis.columns:
#     if "cate" in col.lower():
#         print(col)
        

# df_meslis.mergedLocation_category.unique()

df_meslis.head()

df_meslis['StartTimeIsAssumed']=False

for col in ["StartDate", "StartTime", "EndTime", "EndDate", "LocationDetails", "EventCategory"]:
    df_meslis[col] = None

def define_meslis_source(ad):
    if ad =="":
        return "unknown"
    for link in ["songkick.com", "partyflock","livenation", "eventbrite.com", "setlist.fm", "bandsintown.com", "wegow.com", "concertful.com", "jambase.com", "facebook.com"]:
        if link in ad:
            return link
    return "other"

df_meslis['eventAdvertisements'] = df_meslis['eventAdvertisements'].fillna('')
df_meslis['Source'] = df_meslis['eventAdvertisements'].apply(define_meslis_source)

df_meslis['Source'].value_counts()

flt = df_meslis['Source']=="unknown"
df_meslis.loc[flt]

df_meslis.columns

df_meslis.rename(columns = {"eventAdvertisements": "EventDescription"}, inplace=True)

for col in df_events.columns:
    if col not in df_meslis.columns:
        print(col)

df_meslis['isOnMultipleDays']=None # TODO add this information

for col in df_meslis.columns:
    if col not in df_events.columns:
        # remove column from df_meslis
        df_meslis.drop(columns=[col], inplace=True, errors='ignore')
        


### date field transformations

# data have incorrect time zone - remove them all before transformation

flt = df_meslis.EndDateTime =='0001-01-01T00:00:00Z'
df_meslis.loc[flt, "EndDateTime"]=None

df_meslis.StartDateTime = df_meslis.StartDateTime.astype(str)
df_meslis.EndDateTime = df_meslis.EndDateTime.astype(str)

df_meslis.StartDateTime

for str_part in ["+02:00", "+01:00", "+00:00", "Z"]:
    for colname in ['StartDateTime', "EndDateTime"]:
        df_meslis[colname] = df_meslis[colname].str.replace(str_part, "", regex=False)

df_meslis ["StartDateTime_len"] = df_meslis.StartDateTime.apply(len)
df_meslis ["EndDateTime_len"] = df_meslis.EndDateTime.apply(lambda x: len(x) if x else 0)

df_meslis.StartDateTime_len.value_counts()

df_meslis.EndDateTime_len.value_counts()

df_meslis.EndDateTime = df_meslis.EndDateTime.str.replace("None", "", regex=False)

# add 1 day to end date when the interval is negative 

df_meslis.StartDateTime = pd.to_datetime(df_meslis.StartDateTime, format='ISO8601')

# sorted(df_meslis.EndDateTime.dropna().unique().tolist())

df_meslis.EndDateTime = pd.to_datetime(df_meslis.EndDateTime, format='ISO8601', yearfirst=True)

df_meslis.drop(columns=["StartDateTime_len", "EndDateTime_len"], inplace=True, errors="ignore")


df_meslis['Duration'] =  df_meslis.EndDateTime - df_meslis.StartDateTime

# sorted(df_meslis.Duration.dropna().unique().tolist())

df_meslis.EndDateTime.dt.hour

# if interval is negative, add 1 day to end date
flt = (df_meslis.Duration.dt.days < 0) 
df_meslis.loc[flt, "EndDateTime"] = df_meslis.loc[flt].EndDateTime + pd.Timedelta(days=1)


df_meslis['Duration'] =  df_meslis.EndDateTime - df_meslis.StartDateTime

flt = df_meslis.StartDateTime == df_meslis.EndDateTime
df_meslis.loc[flt, "EndDateTime"] = None

df_meslis['Duration'] =  df_meslis.EndDateTime - df_meslis.StartDateTime

df_meslis['Duration'].value_counts().sort_index()

df_meslis['isOnMultipleDays']=False


# filter duration > 1 days and 0 hours
flt = df_meslis.Duration > pd.Timedelta(days=1)
df_meslis.loc[flt, "isOnMultipleDays"] = True


flt = df_meslis.isOnMultipleDays==False
df_meslis_to_add = df_meslis[flt].copy()

df_meslis_left = df_meslis[~flt]

flt = df_meslis_left.StartDateTime.dt.hour < df_meslis_left.EndDateTime.dt.hour 
print(len(df_meslis_left))
print(len(df_meslis_left[flt]))

df_events.columns

rows_to_append = []
flt = df_meslis_left.StartDateTime.dt.hour < df_meslis_left.EndDateTime.dt.hour 
for i, row in df_meslis_left[flt].iterrows():
    start_date_time = row["StartDateTime"]
    end_date_time = row["EndDateTime"]
    print(f"{start_date_time=}")
    print(f"{end_date_time=}")
    date_range = pd.date_range(start_date_time.date(), end_date_time.date(), freq='D')
    print(f"{date_range=}")
    
    for date in date_range: 
        row_to_append = row.copy()
        row_to_append["StartDate"] = date.strftime('%Y-%m-%d')
        row_to_append["StartDateTime"] = None
        row_to_append["EndDateTime"] = None
        #hour and minute format
        row_to_append["StartTime"] = start_date_time.strftime('%H:%M')
        row_to_append["EndTime"] = end_date_time.strftime('%H:%M')

        rows_to_append.append(row_to_append)

len(rows_to_append)

df_meslis_left = df_meslis_left[~flt]

rows_to_append_2 = []
for i, row in df_meslis_left.sort_values("Duration").iterrows():
    start_date_time = row["StartDateTime"]
    end_date_time = row["EndDateTime"]
    # print(f"{start_date_time=}")
    # print(f"{end_date_time=}")
    date_range = pd.date_range(start_date_time.date(), end_date_time.date(), freq='D')
    print(f"{date_range=}")
    
    for date in date_range: 
        row_to_append = row.copy()
        row_to_append["StartDate"] = date.strftime('%Y-%m-%d')
        if date.date() == start_date_time.date():
            row_to_append["StartTime"] = start_date_time.strftime('%H:%M')
        else:
            row_to_append["StartTime"] = "00:00"

        if date.date() == end_date_time.date(): 
            row_to_append["EndTime"] = end_date_time.strftime('%H:%M')
        else:
            row_to_append["EndTime"] = "23:59"

        rows_to_append_2.append(row_to_append)
    
    print(rows_to_append_2)
    break

# pd.DataFrame(rows_to_append_2)

df_events = pd.concat([df_events, df_meslis_to_add, pd.DataFrame(rows_to_append), pd.DataFrame(rows_to_append_2)], ignore_index=True)

df_events

# flt = df_meslis.Duration.isna()
# df_meslis.loc[~flt].sort_values("Duration").tail(20)


### remove matched meslis events from gema dataset

df_meslis_original = pd.read_pickle(f"{EVENTS_FOLDER}meslis/df_meslis 2024-12-19_12-41-21.pkl")

len(df_meslis_original)

for id in df_meslis_original.id.tail(30):
    print(id,end = " ")

flt = df_events.origin_id.astype(str).str.contains("1006873109")
df_events[flt]

df_meslis_original.mergedLocation_customers.value_counts()

df_subset = df_meslis_original.loc[:,['id', "customers"]].copy()

df_subset.dropna(subset="customers", inplace=True)

flt = df_subset.customers==""
df_subset= df_subset[~flt]

df_subset.customers = df_subset.customers.apply(lambda x: ast.literal_eval(x))

df_subset['customer_len'] = df_subset.customers.apply(len)

df_subset.customer_len.value_counts()

flt = df_subset.customer_len==2
df_subset.loc[flt, "customer_1"] = df_subset.loc[flt].customers.apply(lambda x: x[0])
df_subset.loc[flt, "customer_2"] = df_subset.loc[flt].customers.apply(lambda x: x[0])

df_subset.loc[df_subset.customer_len==2, "customers"]=None

gema_matched_ids_to_exclude = []

gema_matched_ids_to_exclude.extend(df_subset.customers.dropna().apply(lambda dct: dct.get("matchedEventIds")).dropna().tolist())

gema_matched_ids_to_exclude.extend(df_subset.customer_1.dropna().apply(lambda dct: dct.get("matchedEventIds")).dropna().tolist())

gema_matched_ids_to_exclude.extend(df_subset.customer_2.dropna().apply(lambda dct: dct.get("matchedEventIds")).dropna().tolist())

gema_matched_ids_to_exclude = [el for el in gema_matched_ids_to_exclude if el!=[]]

# flatten a list of lists into a single list
gema_matched_ids_to_exclude = [el for sublist in gema_matched_ids_to_exclude for el in sublist]

len(gema_matched_ids_to_exclude)

gema_matched_ids_to_exclude=[str(el) for el in gema_matched_ids_to_exclude]

len(df_events)

flt = df_events.origin_id.astype(str).isin(gema_matched_ids_to_exclude)
print(len(df_events[flt]))
df_events= df_events[~flt]
print(len(df_events))

df_events.to_pickle(f"{EVENTS_FOLDER}/df_events {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.pkl")
