==> notebooks/06c what is the optimal coorinate preicision.ipynb <==
import geopy.distance

lat = 51.01855
lat_plus_delta = 51.01855+0.00001
lng = 13.77314 
lng_plus_delta = lng+0.00001

lat_plus_delta

lng_plus_delta

coords_1 = (lat, lng)
coords_2 = (lat_plus_delta, lng)
geopy.distance.geodesic(coords_1, coords_2).m 

coords_1 = (lat, lng)
coords_2 = (lat, lng_plus_delta)
geopy.distance.geodesic(coords_1, coords_2).m 

lat = 51.0185
lat_plus_delta = lat+0.0001
lng = 13.7731 
lng_plus_delta = lng+0.0001

coords_1 = (lat, lng)
coords_2 = (lat_plus_delta, lng)
geopy.distance.geodesic(coords_1, coords_2).m 

coords_1 = (lat, lng)
coords_2 = (lat, lng_plus_delta)
geopy.distance.geodesic(coords_1, coords_2).m 

lat = 51.018
lat_plus_delta = lat+0.001
lng = 13.773
lng_plus_delta = lng+0.001

coords_1 = (lat, lng)
coords_2 = (lat_plus_delta, lng)
geopy.distance.geodesic(coords_1, coords_2).m 

coords_1 = (lat, lng)
coords_2 = (lat, lng_plus_delta)
geopy.distance.geodesic(coords_1, coords_2).m
