==> notebooks/07b make weather data hourly.ipynb <==
from tqdm.notebook import tqdm
from datetime import datetime
import pandas as pd
import matplotlib.pyplot as plt
import os
import json
import re
import numpy as np
import ast

# Dresden  Station ID = 1048 , Stations-Kennziffer 10488 (Klotzsche)
# Freiburg: Station ID - 1443, 10803  - ? 
 
# 10 minute data https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/10_minutes/

# Columns
# TT_10 temperature
# RF_10 relative humidity 
# RWS_10 Sum of the precipitation height of the previous 10 minutes
# FF_10 mean wind speed during the previous 10 minutes



# Dresden

df_air_DD = pd.read_csv("../data/weather/Dresden/Dresden temperature.txt", sep=";")

df_air_DD.head()

df_air_DD.describe()

df_precip_DD = pd.read_csv("../data/weather/Dresden/Dresden precipitation.txt", sep=";")

df_precip_DD.head()

df_precip_DD.describe()

df_wind_DD = pd.read_csv("../data/weather/Dresden/Dresden wind.txt", sep=";")

df_wind_DD.head()

for df in [df_air_DD, df_precip_DD, df_wind_DD]:
    df.columns = [col.strip() for col in df.columns]
    df["MESS_DATUM"] = pd.to_datetime(df["MESS_DATUM"], format= "%Y%m%d%H%M")
    df.drop(columns=["QN", "eor"], inplace=True, errors="ignore")
    print(len(df))

df_air_DD

df_tmp = pd.merge(df_air_DD, df_precip_DD, on=["MESS_DATUM", "STATIONS_ID"])

df_DD = pd.merge(df_tmp, df_wind_DD, on=["MESS_DATUM", "STATIONS_ID"])

len(df_DD)

assert len(df_DD.STATIONS_ID.unique().tolist())==1

df_DD["MESS_DATUM"].min()

df_DD["MESS_DATUM"].max()

pd.to_datetime("2024-01-01")

DD_dt_flt = ((df_DD["MESS_DATUM"] >= "2024-01-01") & (df_DD["MESS_DATUM"] < "2024-04-01")) | \
 ((df_DD["MESS_DATUM"] >= "2024-09-01") & (df_DD["MESS_DATUM"] < "2024-11-01")) 
len(df_DD[DD_dt_flt])

len(df_DD)

df_DD = df_DD[DD_dt_flt]

df_DD

df_DD = df_DD[["MESS_DATUM", "TT_10", "RF_10", "RWS_10", "FF_10"]]

#df_DD['city']="Dresden"

df_DD.head()

# Freiburg

df_air_recent_FB = pd.read_csv("../data/weather/Freiburg/Freiburg recent temperature.txt", sep=";")

df_air_recent_FB.head()

df_air_hist_FB = pd.read_csv("../data/weather/Freiburg/Freiburg historical air temperature.txt", sep=";")

df_air_FB = pd.concat([df_air_recent_FB, df_air_hist_FB])

df_air_FB = df_air_FB.drop_duplicates()

df_air_FB.head()

df_precip_FB_recent = pd.read_csv("../data/weather/Freiburg/Freiburg recent precipitation.txt", sep=";")

df_precip_FB_hist = pd.read_csv("../data/weather/Freiburg/Freiburg historical precipitation.txt", sep=";")

df_precip_FB = pd.concat([df_precip_FB_recent, df_precip_FB_hist])

len(df_precip_FB)

df_precip_FB = df_precip_FB.drop_duplicates()

df_wind_FB_recent = pd.read_csv("../data/weather/Freiburg/Freiburg recent wind.txt", sep=";")

df_wind_FB_hist = pd.read_csv("../data/weather/Freiburg/Freiburg historical wind.txt", sep=";")

df_wind_FB = pd.concat([df_wind_FB_recent, df_wind_FB_hist])

df_wind_FB = df_wind_FB.drop_duplicates()

for df in [df_air_FB, df_precip_FB, df_wind_FB]:
    df.columns = [col.strip() for col in df.columns]
    df["MESS_DATUM"] = pd.to_datetime(df["MESS_DATUM"], format= "%Y%m%d%H%M")
    df.drop(columns=["QN", "eor"], inplace=True, errors="ignore")
    print(len(df))

df_tmp = pd.merge(df_air_FB, df_precip_FB, on=["MESS_DATUM", "STATIONS_ID"])

df_FB = pd.merge(df_tmp, df_wind_FB, on=["MESS_DATUM", "STATIONS_ID"])

df_FB.head()

assert len(df_FB.STATIONS_ID.unique().tolist())==1

df_FB["MESS_DATUM"].min()

df_FB["MESS_DATUM"].max()

FB_dt_flt = ((df_FB["MESS_DATUM"] >= "2023-06-01") & (df_FB["MESS_DATUM"] < "2023-08-01")) | \
 ((df_FB["MESS_DATUM"] >= "2024-09-01") & (df_FB["MESS_DATUM"] < "2024-11-01")) 
len(df_FB[FB_dt_flt])

len(df_FB)

df_FB = df_FB[FB_dt_flt]

df_FB

df_FB = df_FB[["MESS_DATUM", "TT_10", "RF_10", "RWS_10", "FF_10"]]

df_FB = df_FB.replace(-999, np.nan)

# df_FB['city']="Freiburg"

# unite

df_FB.head()

df_DD.head()

col_renaming = { "MESS_DATUM": "Date", "TT_10": "Temperature", "RF_10": "Humidity", "RWS_10": "Precipitation", "FF_10": "Wind" }

df_DD.rename(columns=col_renaming, inplace=True)

df_FB.rename(columns=col_renaming, inplace=True)

df_DD.set_index("Date", inplace=True)

df_FB.set_index("Date", inplace=True)

len(df_DD)

21888/14640

df_DD.sort_values("Date").head()

grouped_DD = df_DD.resample("h").mean()

grouped_DD.head()

grouped_FB = df_FB.resample("h").mean()

FB_dt_flt = ((grouped_FB.index >= "2023-06-01") & (grouped_FB.index < "2023-08-01")) | \
 ((grouped_FB.index >= "2024-09-01") & (grouped_FB.index < "2024-11-01")) 
len(grouped_FB[FB_dt_flt])

grouped_FB = grouped_FB[FB_dt_flt]

grouped_FB.describe()

grouped_FB[grouped_FB.Precipitation == -999]

assert len(df_FB)/len(grouped_FB)==6

grouped_FB.tail()

len(df_DD.resample("h").mean())

21888/3

df_DD.head()

df_DD.resample("h").mean()

DD_dt_flt = ((grouped_DD.index >= "2024-01-01") ) & (grouped_DD.index < "2024-04-01") | \
 ((grouped_DD.index >= "2024-09-01") & (grouped_DD.index < "2024-11-01")) 
len(grouped_DD[DD_dt_flt])
grouped_DD = grouped_DD[DD_dt_flt]

df_DD

grouped_DD.describe()

grouped_FB.describe()

grouped_FB.info()

df_missing_FB = grouped_FB[grouped_FB.Temperature.isna()].copy()

df_missing_FB

for col in ["Temperature", "Humidity", "Precipitation", "Wind"]:
    grouped_FB[col] = grouped_FB[col].interpolate(method = "nearest")

grouped_FB.loc[df_missing_FB.index]

len(df_FB)/3

grouped_FB.describe()

assert len(df_DD)/len(grouped_DD)==6

time = datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
grouped_DD.to_pickle(f"../data/weather/df_Dresden_weather_hourly {time}.pkl")
grouped_DD.to_csv(f"../data/weather/df_Dresden_weather_hourly {time}.csv")

grouped_FB.to_pickle(f"../data/weather/df_Freiburg_weather_hourly {time}.pkl")
grouped_FB.to_csv(f"../data/weather/df_Freiburg_weather_hourly {time}.csv")
