==> notebooks/08d show bike seasonality.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
import matplotlib.pyplot as plt
import os
import seaborn as sns


CITY = 'DD'

filename_DD = '../data/nextbike/trips_Dresden 2025-01-29_17-48-45.pkl' # 
filename_FB = '../data/nextbike/trips_Freiburg 2025-01-29_17-48-45.pkl' # 
help_dict = {'DD': filename_DD, 'FB': filename_FB}

df = pd.read_pickle(help_dict[CITY])

df = df.loc[df.date_rent == pd.to_datetime("2024-02-02").date()]

df.head()

df['month_rent']= df.datetime_rent.dt.month_name()

df['weekday_rent']  =  df.datetime_rent.dt.day_name()
df['weekday_rent']

df['hour_rent'] = df.datetime_rent.dt.hour

# weekday, month

ntrips = df.groupby(['month_rent', 'weekday_rent']).size()

ndays = df.groupby(['month_rent', 'weekday_rent'])['date_rent'].nunique()

ndays.head(4)

ntrips.head(4)

# ntrips.loc["January", "Friday"]

# ndays.loc["January", "Friday"]

avg_trips= ntrips/ndays

avg_trips = avg_trips.reset_index(name="Count")

avg_trips.head(4)

weekday_order = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"]
month_order = ["January", "February", "March", "April", "May", "June", 
               "July", "August", "September", "October", "November", "December"]
present_month_order = [m for m in month_order if m in avg_trips.month_rent.unique()]

avg_trips["weekday_rent"] = pd.Categorical(avg_trips["weekday_rent"], categories=weekday_order, ordered=True)
avg_trips.month_rent = pd.Categorical(avg_trips["month_rent"], categories=present_month_order, ordered=True)


plt.figure(figsize=(6, 5))
sns.reset_defaults()
sns.set_style("white")
sns.set_palette("husl")

sns.lineplot(data=avg_trips, x='weekday_rent', y='Count', hue='month_rent', style='month_rent')
plt.xlabel('Weekday')
plt.ylabel('Average number of trips per day')
if CITY=='DD':
    anchor=1.3
else:
    anchor=1.05

plt.legend(title="Month", bbox_to_anchor=(anchor, 1)) #plt.grid(True)
if CITY=='DD':
    plt.ylim(bottom=0, top=9000)
else:
    plt.ylim(bottom=0, top=4000)

plt.grid(True) 
plt.savefig(f'/Users/v.sinichenko/Downloads/plots/seasonality_{CITY}_weekday.png', format='png', bbox_inches='tight')
plt.show()

# flt = (df.weekday_rent=='Friday') & (df.month_rent=='January')
# len(df[flt])

flt = (df.weekday_rent=='Friday') & (df.month_rent=='July')
len(df[flt].date_rent.unique())

14404/4

len(df)

len(df.date_return.unique())

len(df)/len(df.date_return.unique())

# hour, weekday

grouping_cols = ['hour_rent', 'weekday_rent']

ntrips = df.groupby(grouping_cols).size()

ntrips

nintervals = df.groupby(grouping_cols)['datetime_hour_rent'].nunique()

nintervals

avg_trips= ntrips/nintervals

avg_trips.head(4)

avg_trips = avg_trips.reset_index(name="Count")

avg_trips

# avg number of trips per hour
len(df)/len(df.datetime_hour_rent.unique())

avg_trips.sort_values(by='Count', ascending=False)

avg_trips["weekday_rent"] = pd.Categorical(avg_trips["weekday_rent"], categories=weekday_order, ordered=True)

avg_trips.loc[avg_trips.weekday_rent=='Wednesday']

plt.figure(figsize=(6, 5))
sns.reset_defaults()
sns.set_style("white")
sns.set_palette("husl")

sns.lineplot(data=avg_trips, x='hour_rent', y='Count', hue='weekday_rent', style='weekday_rent')
plt.ylabel('Average number of trips per hour')
plt.xlabel('Hour')
if CITY=='DD':
    anchor=1.05
else:
    anchor=1.05

plt.legend(title="Weekday", bbox_to_anchor=(anchor, 1)) #plt.grid(True)
if CITY=='DD':
    plt.ylim(bottom=0, top=2000)
else:
    plt.ylim(bottom=0, top=200)

plt.grid(True) 
plt.savefig(f'/Users/v.sinichenko/Downloads/plots/seasonality_{CITY}_hour_weekday.png', format='png', bbox_inches='tight')
plt.show()
