==> notebooks/08e plot daily tip counts.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
import matplotlib.pyplot as plt
import os
import seaborn as sns
import numpy as np


filename_DD = '../data/nextbike/trips_Dresden 2025-01-29_17-48-45.pkl' # 
filename_FB = '../data/nextbike/trips_Freiburg 2025-01-29_17-48-45.pkl' # 


df_DD = pd.read_pickle(filename_DD)

df_FB = pd.read_pickle(filename_FB)

df_DD['City']='Dresden'
df_FB['City']='Freiburg'

df = pd.concat([df_DD,df_FB])

grouped= df.groupby(['City','date_rent']).size().reset_index(name='Count')

grouped

grouped = grouped.sort_values(by=['date_rent'])

min_date, max_date = grouped['date_rent'].min(), grouped['date_rent'].max()

date_range = pd.date_range(start=min_date, end=max_date)

multi_index = pd.MultiIndex.from_product([df["City"].unique(), date_range], names=["City", "date_rent"])


grouped = grouped.set_index(["City", "date_rent"]).reindex(multi_index).reset_index()

grouped.Count.iloc[0]

grouped.Count.fillna(pd.NA).iloc[0]

grouped.Count.astype('Int64').fillna(pd.NA).iloc[0]

grouped.Count =  grouped.Count.astype('Int64').fillna(pd.NA)

grouped.Count.iloc[0]

grouped.loc[(grouped.City=='Dresden') & (grouped.date_rent=='2024-07-01')]

plt.figure(figsize=(20, 10))


import matplotlib.cm as cm

# Create the figure
plt.figure(figsize=(8, 5))

colors = plt.colormaps.get_cmap("Dark2")

# Loop through each city and plot its line separately
for i, city in enumerate(['Dresden', "Freiburg"]):
    city_data = grouped[grouped["City"] == city]  # Filter data for the city
    plt.plot(city_data["date_rent"], city_data["Count"], linestyle="-", label=city,color=colors(i))

# Formatting
plt.xticks(rotation=45)  # Rotate x-axis labels for better readability
plt.ylabel("Number of trips per day")
plt.xlabel("Date")
plt.legend(title="City")
plt.grid(True)
plt.ylim(bottom=0, top=16000)
plt.savefig(f'/Users/v.sinichenko/Downloads/plots/daily_trips.png', format='png', bbox_inches='tight')
plt.show()


grouped = grouped.dropna(subset="Count")

subset_city = grouped.loc[(grouped.City=='Dresden')].sort_values(by='date_rent',ascending=True)

subset = grouped.loc[grouped.City=='Freiburg'].sort_values(by='Count',ascending=False).head(20)

grouped.loc[grouped.City=='Freiburg'].sort_values(by='Count',ascending=False)

subset.sort_values(by='date_rent',ascending=True)
