==> notebooks/08f descriptive stats of events.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
import matplotlib.pyplot as plt
import os
import seaborn as sns
import numpy as np
import plotly.express as px
from datetime import datetime


filename = '../data/events/df_events 2025-03-02_12-26-38.pkl' # 

df = pd.read_pickle(filename)

df.EventCategory.value_counts()

# preprocessing

df.City.value_counts()

df.loc[df.City!="Freiburg", "City"]="Dresden"

df.StartDate = df.StartDateTime.dt.date

df.StartDate 

mapping = {"Gema":"GEMA",
           "facebook.com": "deecooob GmbH",
           "browse_ai": "Self scraped", 
           "Football matches": "Self scraped",
           "eventim.de": "deecooob GmbH",
           "bandsintown.com": "deecooob GmbH",
           "songkick.com": "deecooob GmbH",
           "wegow.com": "deecooob GmbH",
           "setlist.fm": "deecooob GmbH",
          "ticketmaster.de": "deecooob GmbH",
          "concertful.com": "deecooob GmbH",
          "jambase.com": "deecooob GmbH",
          "eventbrite.com": "deecooob GmbH",
          "partyflock": "deecooob GmbH",
          "livenation": "deecooob GmbH"         
}

df['SourceGroup'] = df.Source.map(mapping)

df['SourceGroup'].value_counts()

flt = (df.StartDate > pd.to_datetime('2023-06-25').date()) & (df.StartDate < pd.to_datetime('2024-07-01').date()) & (df.City=="Freiburg")
df[flt]


# remove some of the events with wrong dates
initial_len = len(df)


flt = (df.StartDate > pd.to_datetime('2024-03-31').date()) & (df.StartDate < pd.to_datetime('2024-09-03').date()) & (df.City=="Dresden")
df = df[~flt]
print(initial_len - len(df))
initial_len=len(df)

flt = (df.StartDate > pd.to_datetime('2023-06-25').date()) & (df.StartDate < pd.to_datetime('2023-07-01').date()) & (df.City=="Freiburg")
df = df[~flt]
print(initial_len - len(df))
initial_len=len(df)

flt = (df.StartDate > pd.to_datetime('2023-07-31').date()) & (df.StartDate < pd.to_datetime('2024-09-03').date()) & (df.City=="Freiburg")
df = df[~flt]
print(initial_len - len(df))
initial_len=len(df)

flt = df.StartDate > pd.to_datetime('2024-10-31').date()
df = df[~flt]
print(initial_len - len(df))
initial_len=len(df)

# trips per day

grouped= df.groupby(['City','StartDate']).size().reset_index(name='Count')

grouped = grouped.sort_values(by=['StartDate'])

min_date, max_date = grouped['StartDate'].min(), grouped['StartDate'].max()

date_range = pd.date_range(start=min_date, end=max_date)

multi_index = pd.MultiIndex.from_product([df["City"].unique(), date_range], names=["City", "StartDate"])


grouped = grouped.set_index(["City", "StartDate"]).reindex(multi_index).reset_index()

grouped.Count.iloc[0]

grouped.Count.fillna(pd.NA).iloc[0]

grouped.Count.astype('Int64').fillna(pd.NA).iloc[0]

grouped.Count =  grouped.Count.astype('Int64').fillna(pd.NA)

grouped.Count.iloc[0]

city_data = grouped[(grouped["City"] == "Freiburg") & (grouped["StartDate"] < pd.to_datetime('2023-08-01'))]
city_data["Count"] = city_data["Count"].fillna(0)
plt.bar(city_data["StartDate"], city_data["Count"])
plt.xticks(rotation=45)
plt.ylabel("Count")
plt.xlabel("StartDate")
plt.title("Events in Freiburg before 2024")
plt.show()

# Create the figure
plt.figure(figsize=(8, 5))

colors = plt.colormaps.get_cmap("Dark2")

for i, city in enumerate(['Dresden', "Freiburg"]):
    city_data = grouped[(grouped["City"] == city)]
    plt.plot(city_data["StartDate"], city_data["Count"], linestyle="-", label=city,color=colors(i))

# Formatting
plt.xticks(rotation=45) 
plt.ylabel("Number of events per day")
plt.xlabel("Date")
plt.legend(title="City")
plt.grid(True)
#plt.ylim(bottom=0, top=16000)
plt.savefig(f'/Users/v.sinichenko/Downloads/plots/daily_events.png', format='png', bbox_inches='tight')
plt.show()


# Stacked by month

df["StartMonth"]= df.StartDateTime.dt.to_period("M")

grouped= df.groupby(['City','StartMonth', 'SourceGroup']).size().reset_index(name='Count')

grouped

# min_date, max_date = grouped['StartDateTime'].min(), grouped['StartDateTime'].max()

#multi_index = pd.MultiIndex.from_product([df["City"].unique(), date_range], names=["City", "date_rent"])


# grouped = grouped.set_index(["City", "date_rent"]).reindex(multi_index).reset_index()

# grouped.Count.iloc[0]

# grouped.Count.fillna(pd.NA).iloc[0]

# grouped.Count.astype('Int64').fillna(pd.NA).iloc[0]

# grouped.Count =  grouped.Count.astype('Int64').fillna(pd.NA)

# grouped.Count.iloc[0]

husl_colors = sns.color_palette("husl", 5).as_hex()
husl_colors

grouped['StartMonth'] = grouped['StartMonth'].dt.to_timestamp()

category_order = grouped.SourceGroup.unique().tolist()

grouped.StartMonth.unique()

to_map

for city in ["Dresden", "Freiburg"]:
    grey =200
    rgb_colors = ["rgb(188, 38, 26)", f"rgb({grey}, {grey}, {grey})", "rgb(74, 114, 171)"]
    to_map = grouped[grouped["City"] == city].sort_values(by=['SourceGroup']) 
    to_map.StartMonth = to_map.StartMonth.dt.strftime('%Y-%m').astype(str)

    fig = px.bar(to_map, 
                x="StartMonth", 
                y="Count", color="SourceGroup", 
                barmode="stack",
                color_discrete_sequence=rgb_colors)

    # change xtitle
    fig.update_yaxes(title_text='Number of events per month')

    fig.update_layout(template='simple_white')
    fig.update_layout(
        plot_bgcolor='white',  # Set background to white
        paper_bgcolor='white',  # Set outer background to white
        xaxis=dict(
            showgrid=True,  # Show grid lines
            gridcolor='rgb(211, 211, 211)',
            gridwidth=1,
            linecolor='black'
        ),
        yaxis=dict(
            showgrid=True,  # Show grid lines
            gridcolor='rgb(211, 211, 211)',
            gridwidth=1, 
            showline=True,
        ),
        legend_title="Major source",
        margin=dict(t=0, b=0, l=0, r=0), 
        autosize=True )
    
    fig.write_image(f'/Users/v.sinichenko/Downloads/plots/events_by_source_{city}_events.png')
    

    fig.show()

# Table of small sources

df.columns

df.loc[df.SourceGroup=="deecooob GmbH", ['SourceGroup', "Source"]].value_counts().reset_index(name='Count') #.to_csv('/Users/v.sinichenko/Downloads/plots/source_group_counts.csv', index=False)

df[['SourceGroup', "Source"]].value_counts().sort_index().reset_index(name='Count')

flt = df.Source=="Gema"
df.loc[flt, "Source"]="GEMA"

flt = (df.Source=="Football matches") & (df.City=="Dresden")
df.loc[flt, "Source"]="dynamo-dresden.de"

flt = (df.Source=="Football matches") & (df.City=="Freiburg")
df.loc[flt, "Source"]="scfreiburg.com"

df_browse_ai = pd.read_pickle("../data/events/browse_ai/df_browse_ai 2025-01-02_18-50-53.pkl")

df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("https://", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("www.", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/veranstaltungen/heute/", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/veranstaltungen/morgen/", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/sachsen.html", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/baden-wuerttemberg.html", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/veranstaltungen/wochenende/", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/veranstaltungen-baden-wuerttemberg-qqxx", "")
df_browse_ai.Origin_URL = df_browse_ai.Origin_URL.str.replace("/veranstaltungskalender?page=12", "")

(df_browse_ai.Origin_URL.value_counts(normalize = True))*1920

1125 +717 +59+ 20

len(df)

df.City.value_counts()

EVENTS_FOLDER = "../data/events/"
df.to_pickle(f"{EVENTS_FOLDER}/df_events with full StartDateTime {datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.pkl")

EVENTS_FOLDER = "../data/events/"

df = pd.read_pickle(f"{EVENTS_FOLDER}/df_events with full StartDateTime 2025-03-02_18-01-52.pkl")

df.EndDateTime.isna().value_counts(normalize=True)

df['Duration'] =  df.EndDateTime - df.StartDateTime

df['Duration'].value_counts()
