==> notebooks/08g more descriptive stats.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
import matplotlib.pyplot as plt
import os
import seaborn as sns
import numpy as np
import plotly.express as px
from datetime import datetime


EVENTS_FOLDER = "../data/events/"

df = pd.read_pickle(f"{EVENTS_FOLDER}/df_events with full StartDateTime 2025-03-02_18-01-52.pkl")

df["EndDateTimeMissing"]= df.EndDateTime.isna()
df[['EndDateTimeMissing', "SourceGroup"]].value_counts()

df.EndDateTime.isna().value_counts(normalize=True)

pd.crosstab(df["SourceGroup"], df["EndDate"].isna())

df['Duration'] =  df.EndDateTime - df.StartDateTime

df['Duration'].value_counts()

df.columns

df.Duration.dropna().dt.days.value_counts()

df.EventCategory.value_counts()

tmp  = df.loc[df.Duration.dt.days>0, ['Source', 'Duration']].value_counts()

df_browse_ai = pd.read_pickle("../data/events/browse_ai/df_browse_ai 2024-12-06_14-44-30.pkl")

df_browse_ai.columns

df_browse_ai.dropna(subset="extrMultipleDaysDetails")

df_browse_ai[df_browse_ai.isOnMultipleDays==True]

df[df.origin_id==18582]

df.dropna(subset=["EndDateTime", "StartTime"])

df_weather = pd.read_pickle(f"../data/weather/df_Dresden_weather 2025-01-17_14-52-44.pkl")

df_weather.describe()

df_weather[df_weather.Precipitation>0].sample(10)
