==> notebooks/10a calculate trip distance.ipynb <==
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
from shapely.geometry import Point
import contextily as ctx
from geopy.distance import geodesic
from datetime import datetime
import numpy as np

filename_DD = '../data/nextbike/trips_Dresden 2025-01-26_19-17-13.pkl'
filename_FB = '../data/nextbike/trips_Freiburg 2025-01-26_19-17-13.pkl'
df_DD = pd.read_pickle(filename_DD)
df_FB = pd.read_pickle(filename_FB)

pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", 100)
pd.set_option('display.max_colwidth', 400)

df_DD.head()

coords_1 = tuple(df_DD.loc[2, ["lat_rent", "lng_rent"]])
coords_2 = tuple(df_DD.loc[2, ["lat_return", "lng_return"]])
int(geodesic(coords_1, coords_2).m) # distance fluctuate at 45 meters, 24 m... 

start_time = datetime.now()
df_DD['distance_alt'] = df_DD.apply(lambda x: int(geodesic((x['lat_rent'], x['lng_rent']), (x['lat_return'], x['lng_return'])).m), axis=1)
print(datetime.now() - start_time)


df_DD.drop(columns="distance_alt", inplace=True)

(df_DD['distance_alt']==df_DD['distance'] ).all()

# start_time = datetime.now()
# def calculate_distance_vectorized(lat_rent, lng_rent, lat_return, lng_return):
#     return np.array([
#         geodesic((lat_r, lng_r), (lat_rt, lng_rt)).m
#         for lat_r, lng_r, lat_rt, lng_rt in zip(lat_rent, lng_rent, lat_return, lng_return)
#     ], dtype=int)

# # Apply the vectorized function to the DataFrame
# df_DD['distance'] = calculate_distance_vectorized(
#     df_DD['lat_rent'].to_numpy(),
#     df_DD['lng_rent'].to_numpy(),
#     df_DD['lat_return'].to_numpy(),
#     df_DD['lng_return'].to_numpy()
# )
# end_time = datetime.now()

end_time-start_time

start_time = datetime.now()
df_FB['distance'] = df_FB.apply(lambda x: int(geodesic((x['lat_rent'], x['lng_rent']), (x['lat_return'], x['lng_return'])).m), axis=1)
print(datetime.now() - start_time)


time = datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
df_DD.to_pickle(f'../data/nextbike/trips_Dresden {time}.pkl')
df_FB.to_pickle(f'../data/nextbike/trips_Freiburg {time}.pkl')
