==> notebooks/12a SARIMAX for one hexagon (old).ipynb <==
import pmdarima as pm
#import auto_arima # requires numpy version 1.26.4 i.e. before 2. 
import time
import pandas as pd

datetime = "2025-03-19_10-47-56"
filename_DD = f'../data/nextbike/hourly_demand_supply_Dresden {datetime}.csv'
filename_FB = f'../data/nextbike/hourly_demand_supply_Freiburg {datetime}.csv'
df_DD = pd.read_csv(filename_DD, index_col=None, parse_dates=['datetime_hour'])
df_FB = pd.read_csv(filename_FB, index_col=None, parse_dates=['datetime_hour'])

mycell='871f1b559ffffff'

mycell = "871f1b54bffffff"

df_DD

df_DD.groupby("hex_id").agg({'rent_count': 'sum', 
                                'return_count': 'sum'})
                            

df_DD.rent_count.sum()

df_DD.hex_id.unique()

df_FB.hex_id.unique()

df_DD.loc[df_DD.hex_id==mycell, ['datetime_hour', 'rent_count', 'return_count']].plot(x='datetime_hour', y=['rent_count', 'return_count'])

test_range_1_DD = pd.date_range(start='2024-03-21', end='2024-03-31')
test_range_1_DD = [date.date() for date in test_range_1_DD]

test_range_2_DD = pd.date_range(start='2024-10-21', end='2024-10-31')
test_range_2_DD = [date.date() for date in test_range_2_DD]


df_DD_1 = df_DD.loc[df_DD.datetime_hour.dt.date <= test_range_1_DD[-1]]

# df_DD_1.plot(x='datetime_hour', y=['rent_count', 'return_count'])

df_DD_2 = df_DD.loc[df_DD.datetime_hour.dt.date > test_range_1_DD[-1]]

# df_DD_2.plot(x='datetime_hour', y=['rent_count', 'return_count'])

# df_DD_1_for_acf = df_DD_1.groupby('datetime_hour').agg({'rent_count': 'sum', 'return_count': 'sum'})

# df_DD_2_for_acf = df_DD_2.groupby('datetime_hour').agg({'rent_count': 'sum', 'return_count': 'sum'})

# df_DD_for_acf = df_DD.groupby('datetime_hour').agg({'rent_count': 'sum', 'return_count': 'sum'})

# pm.utils.acf(df_DD_for_acf.rent_count, nlags=25)

# df_FB_for_acf = df_FB.groupby('datetime_hour').agg({'rent_count': 'sum', 'return_count': 'sum'})

# pm.tsdisplay(df_DD_2_for_acf.rent_count, lag_max=24*3, title="Dresden", show=True)

# pm.tsdisplay(df_DD_for_acf.return_count, lag_max=24*3, title="Dresden bike sharing supply", show=True)
# pm.tsdisplay(df_FB_for_acf.return_count, lag_max=24*3, title="Freiburg bike sharing supply", show=True)

# pm.tsdisplay(df_DD_for_acf.rent_count, lag_max=24*3, title="Dresden bike sharing demand", show=True)
# pm.tsdisplay(df_FB_for_acf.rent_count, lag_max=24*3, title="Freiburg bike sharing demand", show=True)

# pm.tsdisplay(df_DD_1_for_acf.rent_count, lag_max=24*3, title="Dresden", show=True)

# df_DD.loc[df_DD.datetime_hour.dt.date.isin(test_range_DD) & (df_DD.hex_id==mycell)].sort_values('datetime_hour')

train_validation_DD_1 = df_DD_1.loc[~df_DD_1.datetime_hour.dt.date.isin(test_range_1_DD)]

test_DD_1 = df_DD.loc[df_DD.datetime_hour.dt.date.isin(test_range_1_DD)].sort_values('datetime_hour')

train_validation_DD_2 = df_DD_2.loc[~df_DD_2.datetime_hour.dt.date.isin(test_range_2_DD)].sort_values('datetime_hour')

test_DD_2 = df_DD_2.loc[df_DD_2.datetime_hour.dt.date.isin(test_range_2_DD)].sort_values('datetime_hour')

len(test_DD_2)

len(train_validation_DD_2)

train_1 = train_validation_DD_1[train_validation_DD_1.hex_id==mycell].set_index('datetime_hour')['rent_count']

test_1 = test_DD_1[test_DD_1.hex_id==mycell].set_index('datetime_hour')['rent_count']

train_1 = train_1.asfreq('h')

# test_1.asfreq('h').isna().value_counts()

test_1.asfreq('h').isna().value_counts()

test_1 = test_1.asfreq('h')

start_train_time = time.time()
model = pm.ARIMA(order=(0, 1, 1), seasonal_order=(0, 0, 2, 24))
model.fit(train_1)

print(f"Elapsed time: {(time.time() - start_train_time)/60} minutes")

model.summary()

predictions = model.predict(n_periods=264)

predictions

test_1

from matplotlib import pyplot as plt

x = test_1.index
plt.plot(x, test_1, color='black')
plt.scatter(x, predictions, color='black')
plt.show()
